HIVE-29117: Refactor deleteDir and moveToTrash
